Mental_health_bot/requirements.txt
flask
langchain
langchain_groq
langchain-community
langchain-core
langchain-text-splitters
chromadb
sentence-transformers
python-dotenv
pyttsx3
